@@ -62,68 +63,178 @@
   },
   "editions": [
     {
-      "description": "Source + ESNext + Require",
-      "entry": "index.js",
+      "description": "typescript source code with import for modules",
       "directory": "source",
-      "syntaxes": [
-        "javascript",
-        "esnext",
-        "require",
-        "const",
-        "let",
-        "defaults",
-        "rest",
-        "spread"
-      ]
+      "entry": "index.ts",
+      "tags": [
+        "typescript",
+        "import"
+      ],
+      "engines": false
     },
     {
-      "description": "Babel Compiled + ES2015 + Require",
+      "description": "typescript compiled for browsers with require for modules",
+      "directory": "edition-browsers",
       "entry": "index.js",
-      "directory": "es2015",
-      "syntaxes": [
+      "tags": [
         "javascript",
-        "es2015",
         "require"
-      ]
+      ],
+      "engines": {
+        "node": false,
+        "browsers": "defaults"
+      }
+    },
+    {
+      "description": "typescript compiled for node.js 11 with require for modules",
+      "directory": "edition-node-11",
+      "entry": "index.js",
+      "tags": [
+        "javascript",
+        "require"
+      ],
+      "engines": {
+        "node": "6 || 8 || 10 || 11",
+        "browsers": false
+      }
+    },
+    {
+      "description": "typescript compiled for node.js 0.12 with require for modules",
+      "directory": "edition-node-0.12",
+      "entry": "index.js",
+      "tags": [
+        "javascript",
+        "require"
+      ],
+      "engines": {
+        "node": "0.12 || 4 || 6 || 8 || 10 || 11",
+        "browsers": false
+      }
     }
   ],
+  "types": "source/index.ts",
   "main": "index.js",
-  "browser": "es2015/index.js",
+  "browser": "edition-browsers/index.js",
   "dependencies": {
     "ansicolors": "^0.3.2",
-    "diff": "^3.2.0",
-    "editions": "^1.3.3"
+    "diff": "^3.5.0",
+    "editions": "^2.1.0"
   },
   "devDependencies": {
-    "assert-helpers": "^4.5.0",
-    "babel-cli": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "documentation": "^5.3.5",
-    "eslint": "^4.16.0",
-    "joe": "^2.0.2",
-    "joe-reporter-console": "^2.0.1",
-    "projectz": "^1.4.0",
-    "surge": "^0.19.0"
+    "@babel/cli": "^7.1.5",
+    "@babel/core": "^7.1.6",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-typescript": "^7.1.0",
+    "@types/ansicolors": "0.0.30",
+    "@types/diff": "^3.5.2",
+    "@types/node": "^10.12.11",
+    "babel-plugin-add-module-exports": "^1.0.0",
+    "eslint": "^5.9.0",
+    "eslint-config-bevry": "^1.0.2",
+    "eslint-config-prettier": "^3.3.0",
+    "eslint-plugin-prettier": "^3.0.0",
+    "eslint-plugin-typescript": "^0.14.0",
+    "kava": "^3.1.0",
+    "prettier": "^1.15.3",
+    "projectz": "^1.5.3",
+    "surge": "^0.20.1",
+    "typedoc": "^0.13.0",
+    "typescript": "^3.2.1",
+    "typescript-eslint-parser": "^21.0.1",
+    "valid-directory": "^1.0.0"
   },
   "scripts": {
+    "our:clean": "rm -Rf ./docs ./edition* ./es2015 ./es5 ./out",
+    "our:compile": "npm run our:compile:edition-browsers && npm run our:compile:edition-node-0.12 && npm run our:compile:edition-node-11",
+    "our:compile:edition-browsers": "env BABEL_ENV=edition-browsers babel --extensions \".ts,.tsx\" --out-dir ./edition-browsers ./source",
+    "our:compile:edition-node-0.12": "env BABEL_ENV=edition-node-0.12 babel --extensions \".ts,.tsx\" --out-dir ./edition-node-0.12 ./source",
+    "our:compile:edition-node-11": "env BABEL_ENV=edition-node-11 babel --extensions \".ts,.tsx\" --out-dir ./edition-node-11 ./source",
+    "our:deploy": "echo no need for this project",
+    "our:meta": "npm run our:meta:docs && npm run our:meta:projectz",
+    "our:meta:docs": "typedoc --mode file --exclude '**/*test*' --name \"$npm_package_name\" --readme ./README.md --out ./docs ./source",
+    "our:meta:projectz": "projectz compile",
+    "our:release": "npm run our:release:prepare && npm run our:release:check-changelog && npm run our:release:check-dirty && npm run our:release:tag && npm run our:release:push",
+    "our:release:check-changelog": "cat ./HISTORY.md | grep v$npm_package_version || (echo add a changelog entry for v$npm_package_version && exit -1)",
+    "our:release:check-dirty": "git diff --exit-code",
+    "our:release:prepare": "npm run our:clean && npm run our:compile && npm run our:test && npm run our:meta",
+    "our:release:push": "git push origin master && git push origin --tags",
+    "our:release:tag": "export MESSAGE=$(cat ./HISTORY.md | sed -n \"/## v$npm_package_version/,/##/p\" | sed 's/## //' | awk 'NR>1{print buf}{buf = $0}') && test \"$MESSAGE\" || (echo 'proper changelog entry not found' && exit -1) && git tag v$npm_package_version -am \"$MESSAGE\"",
     "our:setup": "npm run our:setup:npm",
     "our:setup:npm": "npm install",
-    "our:clean": "rm -Rf ./docs ./es2015 ./es5 ./out",
-    "our:compile": "npm run our:compile:es2015",
-    "our:compile:es2015": "babel ./source --out-dir ./es2015 --presets es2015",
-    "our:meta": "npm run our:meta:docs && npm run our:meta:projectz",
-    "our:meta:docs": "documentation build -f html -o ./docs -g --shallow ./source/**.js",
-    "our:meta:projectz": "projectz compile",
-    "our:verify": "npm run our:verify:eslint",
-    "our:verify:eslint": "eslint --fix ./source",
     "our:test": "npm run our:verify && npm test",
-    "our:release": "npm run our:release:prepare && npm run our:release:check && npm run our:release:tag && npm run our:release:push",
-    "our:release:prepare": "npm run our:clean && npm run our:compile && npm run our:test && npm run our:meta",
-    "our:release:check": "npm run our:release:check:changelog && npm run our:release:check:dirty",
-    "our:release:check:changelog": "cat ./HISTORY.md | grep v$npm_package_version || (echo add a changelog entry for v$npm_package_version && exit -1)",
-    "our:release:check:dirty": "git diff --exit-code",
-    "our:release:tag": "export MESSAGE=$(cat ./HISTORY.md | sed -n \"/## v$npm_package_version/,/##/p\" | sed 's/## //' | awk 'NR>1{print buf}{buf = $0}') && test \"$MESSAGE\" || (echo 'proper changelog entry not found' && exit -1) && git tag v$npm_package_version -am \"$MESSAGE\"",
-    "our:release:push": "git push origin master && git push origin --tags",
-    "test": "node --harmony ./test.js --joe-reporter=console"
+    "our:verify": "npm run our:verify:directory && npm run our:verify:eslint && npm run our:verify:typescript",
+    "our:verify:directory": "npx valid-directory",
+    "our:verify:eslint": "eslint --fix --ignore-pattern '**/*.d.ts' './source/**/*.*'",
+    "our:verify:typescript": "tsc --project .",
+    "test": "node ./test.js"
+  },
+  "eslintConfig": {
+    "extends": [
+      "bevry"
+    ]
+  },
+  "prettier": {
+    "semi": false,
+    "singleQuote": true
+  },
+  "babel": {
+    "env": {
+      "edition-browsers": {
+        "presets": [
+          [
+            "@babel/preset-env",
+            {
+              "targets": {
+                "browsers": "defaults"
+              }
+            }
+          ],
+          "@babel/preset-typescript"
+        ],
+        "plugins": [
+          "@babel/proposal-class-properties",
+          "@babel/proposal-object-rest-spread",
+          "add-module-exports"
+        ]
+      },
+      "edition-node-11": {
+        "presets": [
+          [
+            "@babel/preset-env",
+            {
+              "targets": {
+                "node": "11"
+              }
+            }
+          ],
+          "@babel/preset-typescript"
+        ],
+        "plugins": [
+          "@babel/proposal-class-properties",
+          "@babel/proposal-object-rest-spread",
+          "add-module-exports"
+        ]
+      },
+      "edition-node-0.12": {
+        "presets": [
+          [
+            "@babel/preset-env",
+            {
+              "targets": {
+                "node": "0.12"
+              }
+            }
+          ],
+          "@babel/preset-typescript"
+        ],
+        "plugins": [
+          "@babel/proposal-class-properties",
+          "@babel/proposal-object-rest-spread",
+          "add-module-exports"
+        ]
+      }
+    }
   }
 }
