@@ -1,5 +1,4 @@
-// 2016 December 29
-// https://github.com/bevry/editions
 'use strict'
 
+/** @type {typeof import("./source/test.ts") } */
 module.exports = require('editions').requirePackage(__dirname, require, 'test.js')
